Fixed wrong adapter name in the generated index.tsx.
